falta poco

+++ src/main/java/com/alejandro/crud/repository/InventarioRepository.java
@@ -4,5 +4,5 @@ import com.alejandro.crud.entity.Inventario;
 import com.alejandro.crud.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventarioRepository extends JpaRepository<Inventario, Integer> {
+public interface InventarioRepository extends JpaRepository<Inventario, Long> {
 }
