+++ src/main/java/com/alejandro/crud/repository/ServicioRepository.java
@@ -4,5 +4,5 @@ import com.alejandro.crud.entity.Producto;
 import com.alejandro.crud.entity.Servicio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicioRepository extends JpaRepository<Servicio, Integer> {
+public interface ServicioRepository extends JpaRepository<Servicio, Long> {
 }
